requirements.txt
opencv-python
pyserial
Pillow